+++ watchdog_server/src/main/java/com/siemt3/watchdog_server/cep/listener/port/PortDistBlockScanListener.java
@@ -26,8 +26,8 @@ public class PortDistBlockScanListener implements UpdateListener {
 		long arrivalTime = (long) newEvents[0].get("arrivalTime");
 		Integer maxPort = (int) newEvents[0].get("maxPort");
 		Integer minPort = (int) newEvents[0].get("minPort");
-		Integer dstIpCount = (int) (long) newEvents[0].get("dstIpCount");
-		Integer srcIpCount = (int) (long) newEvents[0].get("srcIpCount");
+		Integer dstIpCount = (int) newEvents[0].get("dstIpCount");
+		Integer srcIpCount = (int) newEvents[0].get("srcIpCount");
 
 		PortDistBlockScanEvent portDistBlock = new PortDistBlockScanEvent(arrivalTime, maxPort, minPort, dstIpCount,
 				srcIpCount);
